refactor(server): user 와 customer 의 분리 #213
