+++ apps/server/src/services/customer.service.ts
@@ -1,21 +1,157 @@
 import { prisma } from '@/lib/prisma';
 import { AppError } from '@/types';
+import { Customer } from '@prisma/client';
+import { hashPassword } from './auth.service';
 
-/** 기본 주소 변경 */
-export const updateDefaultAddressNo = async (customerNo: number, addressNo: number) => {
-  const updatedCustomer = await prisma.customer.update({
-    where: { no: customerNo },
-    data: {
-      defaultAddressNo: addressNo,
-    },
-  });
-  return updatedCustomer;
-};
-
-/** 고객 조회 */
-export const getOne = async (id: string) => {
+/**
+ * 로그인하기 위한 사용자 정보 조회 (민감정보)
+ * @param id
+ * @returns customer: Customer 테이블 정보
+ */
+export const getByIdForLogin = async (id: string) => {
   const customer = await prisma.customer.findUnique({
     where: { id },
+    select: {
+      no: true,
+      id: true,
+      name: true,
+      pw: true,
+    },
+  });
+
+  return customer;
+};
+
+/** refreshToken 업데이트 (고객) */
+export const updateRefreshToken = async (id: string, refreshToken: string | null) => {
+  const customer = await prisma.customer.update({
+    where: { id },
+    data: { refreshToken },
+  });
+
+  return customer;
+};
+
+/**  Customer 등록 타입 */
+export type CreateCustomerInput = Pick<
+  Customer,
+  | 'id'
+  | 'name'
+  | 'pw'
+  | 'mobileNumber'
+  | 'isServiceTermsAgreed'
+  | 'isPrivacyTermsAgreed'
+  | 'isMarketingTermsAgreed'
+> & {
+  address: string;
+  addressDetail: string;
+  zipcode: string;
+};
+
+/**
+ * 회원가입 : 고객, 배송지, 고객-쿠폰 저장
+ */
+export const createCustomerInfo = async (data: CreateCustomerInput) => {
+  try {
+    const result = await prisma.$transaction(async (tx) => {
+      const {
+        id,
+        name,
+        pw,
+        mobileNumber,
+        address,
+        addressDetail,
+        zipcode,
+        isServiceTermsAgreed,
+        isPrivacyTermsAgreed,
+        isMarketingTermsAgreed,
+      } = data;
+
+      const hashedPw = await hashPassword(pw);
+
+      // 고객 생성
+      const createCustomer = await tx.customer.create({
+        data: {
+          id,
+          name,
+          pw: hashedPw,
+          mobileNumber,
+          isServiceTermsAgreed,
+          isPrivacyTermsAgreed,
+          isMarketingTermsAgreed,
+        },
+      });
+
+      if (createCustomer) {
+        // 배송지 저장
+        const addressInfo = await tx.address.create({
+          data: {
+            address,
+            addressDetail,
+            zipcode,
+            message: '문 앞에 놓아주세요.',
+            recipientName: name,
+            recipientMobile: mobileNumber,
+            customerNo: createCustomer.no,
+          },
+        });
+
+        // 고객-기본배송지 연결
+        await tx.customer.update({
+          where: { no: createCustomer.no },
+          data: {
+            defaultAddressNo: addressInfo.no,
+          },
+        });
+
+        // 쿠폰 조회
+        const coupon = await tx.coupon.findFirst({
+          where: { no: 1 },
+          select: {
+            no: true,
+            expireAfterDays: true,
+          },
+        });
+
+        if (coupon) {
+          const date = new Date();
+          // 쿠폰 만료일 = 현재(발생)일자 + 발급일 기준 만료일
+          date.setDate(date.getDate() + coupon.expireAfterDays);
+          const expiredAt = date;
+
+          // 고객-쿠폰 발급정보 저장
+          await tx.customerCoupon.create({
+            data: {
+              issuedAt: new Date(),
+              expiredAt,
+              isUsed: false,
+              isExpired: false,
+              couponNo: coupon.no,
+              customerNo: createCustomer.no,
+            },
+          });
+        }
+      }
+
+      return createCustomer;
+    });
+
+    return result;
+  } catch (e) {
+    throw AppError.internalServerError(
+      '고객 정보를 등록하는 과정에서 문제가 발생했습니다. 관리자 확인이 필요합니다. (info)',
+    );
+  }
+};
+
+/** 존재하는지만 확인하는 조회 함수 */
+export const getOneForCheck = async (id: string) => {
+  const customer = await prisma.customer.findUnique({
+    where: { id },
+    select: {
+      no: true,
+      id: true,
+    },
   });
 
   if (!customer) {
@@ -23,20 +159,4 @@ export const getOne = async (id: string) => {
   }
 
   return customer;
-};
-
-/**
- * refreshToken 초기화 (고객)
- */
-export const invalidateRefreshToken = async (no: number, id: string) => {
-  const result = await prisma.$transaction(async (tx) => {
-    const update = await tx.customer.update({
-      where: { no, id },
-      data: { refreshToken: null },
-    });
-
-    return update;
-  });
-
-  return result;
 };
